QGraphicsView: replace some Q_FOREACH loops over const locals with C++11 range-for
This needs to be handled a bit carefully, because Qt containers
will detach upon being iterated over using range-for.

In the cases of this patch, that cannot happen, because all
containers are local and marked as const (either by this patch
or before).

Separate patches will deal with other situations.

Range-for loops are much more efficient than foreach loops.

This patch shaves ~1.8KiB of text size off an optimized Linux
AMD64 GCC 4.9 build.

Change-Id: I5c58658937ac4323594161bf94a2fce3c5667914
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>