QAbstractItemView - Fix docs to match behavior (SelectionMode)
We can deselect the item in single selection mode.
It was introduced in
28a21d98ef8d880a6dd86ee19dd803424bb5eae1

The motivation was QTBUG-8836 - and furthermore it was in
many situations already possible to deselect the item by
clicking in the viewport.

A restore of selection by reacting on the signal
QItemSelectionModel::selectionChanged is easily possible

If the behavior was changed to follow the documentation,
such a similar simply work-around to allow deselect would not be
possible.

Task-number: QTBUG-49217
Change-Id: I5466337dc547414d3e0c045ab35d618dba56c508
Reviewed-by: David Faure <david.faure@kdab.com>