Raster and Blitter engines: method override markups and fixes
While investigating a rendering bug I noticed that methods marked in the
blitter engine as virtual and intended to override a base implementation
where in fact just shadowing a non-virtual method in the raster engine.

Fix that and add Q_DECL_OVERRIDE all around to avoid similar breakage
in the future.

Change-Id: I24c5bb27fb6d3ac57dddb704a14dab7ce153a72a
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>