QLineEdit: hold SideWidgetEntry in std::vector, not QVector
This is private implementation, so there's no BC issue here.
The collections ported here also do not benefit from CoW,
because they are never copied.

Adapt to STL API and replace foreach with C++11 range-for
loops, because the former deep-copies STL containers.

Also replace index-based for loops with C++11 range-for,
to evade the int/size_t problem on MSVC.

Saves a bit more than 1KiB in text size on optimized GCC 4.9
Linux AMD64 builds, not all of which can be attributed to the
ports to range-for.

Change-Id: I240030180bd1b2ca40c002b03ab72319a99a87c3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>