Remove a use of a QMap in QAccessibleTextWidget::attributes()
The QMap<QByteArray, QString> was only used to hold key/value pairs
until they were serialized into a QString at the end of the function.

Serialize into QString directly instead, dropping the temporary QMap.

The problem is that we'd like to keep the line formatting central
while at the same time harnessing the power of QStringBuilder.

This is possible with a templated function with an input/output
parameter:

     format_attr(QString &result, const char *key, T &&value)
     format_attr(result, text-foo, str + str2 + str3);

Or with return type deduction:

     formatted(const char *key, T &&value) -> decltype((expr)) { return expr; }
     result += formatted(text-foo, str + str2 + str3);

I don't like out parameters, and we can't rely on auto return
type deduction, yet, so I opted for a miniature expression
template solution that can only match the expression

     attr[key] = value;

where 'key' is a const char* and 'value' can be anything that
QStringBuilder supports. This allows to keep the syntax of a map
while at the same time serializing to QString immediately.

The only behavioral difference to the old code is that the
attributes are no longer sorted, but order doesn't matter.

Saves more than 10KiB in text size on optimized GCC 4.9 Linux
AMD64 builds.

Change-Id: I7b3bec0466ef24156c693adaa95f0316007e0bfe
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>